DAC6-3980: Scala 3 update
Updated scala version to 3

+++ app/config/AppConfig.scala
@@ -46,8 +46,8 @@ class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig
 
   lazy val sendEmailUrl: String = servicesConfig.baseUrl("email")
 
-  lazy val cacheTtl: Int      = config.get[Int]("mongodb.timeToLiveInSeconds")
-  lazy val submissionTtl: Int = config.get[Int]("mongodb.submission.timeToLiveInDays")
+  lazy val cacheTtl: Long      = config.get[Long]("mongodb.timeToLiveInSeconds")
+  lazy val submissionTtl: Long = config.get[Long]("mongodb.submission.timeToLiveInDays")
 
   lazy val emailOrganisationSuccessfulTemplate: String   = config.get[String]("emailTemplates.fileOrganisationUploadSuccessful")
   lazy val emailOrganisationUnsuccessfulTemplate: String = config.get[String]("emailTemplates.fileOrganisationUploadUnsuccessful")
